Rename test file
