+++ protocol/clproto_cpp/test/tests/test_cartesian_space_proto.cpp
@@ -8,7 +8,7 @@
 #include <state_representation/space/cartesian/CartesianWrench.hpp>
 
 #include "clproto.h"
-#include "encode_decode_helper.hpp"
+#include "test_encode_decode.hpp"
 
 using namespace state_representation;
