+++ protocol/clproto_cpp/test/tests/test_joint_space_proto.cpp
@@ -8,7 +8,7 @@
 #include <state_representation/space/joint/JointTorques.hpp>
 
 #include "clproto.h"
-#include "encode_decode_helper.hpp"
+#include "test_encode_decode.hpp"
 
 using namespace state_representation;
