+++ protocol/clproto_cpp/test/tests/test_messages.cpp
@@ -6,7 +6,7 @@
 #include <state_representation/space/cartesian/CartesianPose.hpp>
 
 #include "clproto.h"
-#include "encode_decode_helper.hpp"
+#include "test_encode_decode.hpp"
 
 using namespace state_representation;
